fix(test): 웹폰트 커버리지 가드가 정규식이 아니라 함수를 부르게 한다
token-coverage.test.ts 의 다섯 번째 어서션이 findFontDisplaySrc 를 부르지 않고
원문에 /^\s*font-\w+-src:/m 을 걸어 세고 있었다. 그래서 그 함수가 마이그레이션
후 항상 null 을 반환하게 돼도 테스트는 통과한다 — frontmatter 로 옮겨간 줄을
정규식은 여전히 세기 때문이다.

이 파일을 만든 이유가 "정규식이 조용히 0이 되는 것을 막자" 였는데 가드 자체가
같은 함정에 있었다. 실측으로 확인했다.

  현행 본문      findFontDisplaySrc → 해결됨
  이전 후 모사   findFontDisplaySrc → null    (새 가드가 실패시킨다)
                 원문 정규식        → 3건     (옛 가드는 통과했을 것)

preview-validator 의 findFontDisplaySrc 를 export 한다. 테스트가 실제 소비
경로를 부르지 않으면 그 경로가 끊긴 것을 알 수 없다.

fenceRows 도 token-migration-rules 로 옮겨 스크립트와 테스트가 한 구현을
공유하게 한다. 두 벌로 두면 블록 스칼라 처리가 갈리고, 갈린 쪽이 값을 잃는다.
