+++ src/lib/token-migration-rules.ts
@@ -155,6 +155,66 @@ export const DIRECTIVES: Record<string, Record<string, Destination>> = {
   },
 }
 
+const TOKEN_SECTIONS = new Set(["Colors", "Typography", "Spacing", "Rounded"])
+
+/**
+ * Fence rows from the four token sections, block scalars included.
+ *
+ * `font-sans: >` carries its value on the following indented lines. Reading
+ * only the marker line loses the font stack entirely — which is exactly the
+ * kind of silent deletion this module exists to prevent. Nine entries write
+ * their font stacks that way.
+ */
+export function fenceRows(body: string): Array<FenceRow> {
+  const out: Array<FenceRow> = []
+  const lines = body.split("\n")
+  let section: string | null = null
+  let fence: string | null = null
+
+  for (let i = 0; i < lines.length; i++) {
+    const line = lines[i]
+    const heading = line.match(/^##\s+(.*)$/)
+    if (heading) {
+      section = heading[1].trim()
+      continue
+    }
+    const marker = line.match(/^```(\w*)/)
+    if (marker) {
+      fence = fence === null ? marker[1] || "none" : null
+      continue
+    }
+    if (fence !== "yaml" || !section || !TOKEN_SECTIONS.has(section)) continue
+
+    const m = line.match(/^([A-Za-z][\w-]*)\s*:\s*(.*)$/)
+    if (!m) continue
+
+    let value = m[2].trim()
+    let note: string | undefined
+    if (value === ">" || value === "|") {
+      const parts: Array<string> = []
+      let j = i + 1
+      while (
+        j < lines.length &&
+        /^\s+\S/.test(lines[j]) &&
+        !/^```/.test(lines[j])
+      ) {
+        parts.push(lines[j].trim())
+        j++
+      }
+      value = parts.join(" ")
+      i = j - 1
+    } else {
+      const comment = value.match(/\s+#\s?(.*)$/)
+      if (comment) {
+        note = comment[1].trim()
+        value = value.slice(0, comment.index).trim()
+      }
+    }
+    out.push({ key: m[1], value, note, section })
+  }
+  return out
+}
+
 const GRADIENT = /^"?(?:linear|radial|conic)-gradient\(/
 const BARE_NUMBER = /^-?\d*\.?\d+$/
 const URL_VALUE = /^https?:\/\//
